--- src/components/apresentacao/style.css
@@ -1,36 +1,0 @@
-.container{
-  width:50%;
-  background:#000036;
-  display:flex;
-  flex-direction:column;
-  align-items:center;
-  padding-bottom:1%;
-}
-.container p{
-  width:65%;
-  font-size:1.2vw;
-  margin: 7% auto ;
-  color: #ffffff;
-  text-align:center;
-}
-.container h3{
-  color: #ffffff;
-  font-size: 2vw;
-  margin:5% 0;
-}
-.container .imagem-apresentacao{
-  margin:0 auto;
-  text-align:center;
-}
-.container .imagem-apresentacao img{
-  width:70%;
-}
-
-@media (max-width:1100px){
-  .container{
-    width:100%;
-  }
-  .container .imagem-apresentacao{
-    display:none;
-  }
-}
