--- src/components/button/style.css
@@ -1,23 +1,0 @@
-.btn-primary{
-  background-color: #ff6b00;
-  color: #fff;
-  padding: 1.2% 11%;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 1.3vw;
-  transition: background-color 0.3s;
-}
-
-Link{
-  color:#FFFFFF;
-  font-weight: 700;
-}
-
-@media (max-width:1100px){
-  .btn-primary{ 
-    font-size:2.8vw;
-    margin-bottom:5%;
-  }
-
-}
