--- src/components/form/style.css
@@ -1,130 +1,0 @@
-/* Estilos globais */
-body {
-  font-family: Arial, sans-serif;
-  background-color: #ffffff;
-  color: #333333;
-  padding: 2rem;
-}
-a{
-  text-decoration: none;
-  color: #111111;
-}
-.form-container{
-  width: 50%;
-}
-.form-container .termos{
-  text-align: center;
-}
-.form-container p{
-  text-align:start;
-  color: #ff6b6b;
-}
-.form-container p span{
-  color:#11111182
-}
-.form-container h2{
- font-size:2vw;
- text-align:center;
-}
-.form-container hr{
-  width:70%;
-  text-align:center;
-}
-
-/* Estilos do formulário */
-form {
-  padding: 2% 10%;
-  display:flex;
-  flex-wrap: wrap;
-  background:#ffffff;
-}
-
-label {
-  display: block;
-  margin-bottom: 0.8rem;
-  font-weight: bold;
-  background:#ffffff;
-  margin:4%;
-}
-form div{
-  min-width: 50%;
-  margin-bottom:1%
-}
-input {
-  width:80%;
-  padding:1rem;
-  border: 1px solid #ccc;
-  border-radius: 10px;
-  background-color: #f5f5f5;
-  color: #333;
-  cursor: pointer;
-}
-form .btn-submit{
-  width: 100%;
-  text-align: center;
-}
-.btn-btn-submit {
-  background-color: #ff6b00;
-  color: #fff;
-  padding: 1% 8%;
-  border: none;
-  border-radius: 10px;
-  cursor: pointer;
-  font-size: 1vw;
-  transition: background-color 0.3s;
-} */
-
-.btn-btn-submit:hover {
-  background-color: #e65c00;
-}
-
-p.error {
-  color: #ff6b6b;
-  margin-top: 0.5rem;
-}
-
-@media (max-width:1100px){
-  .container p {
-    font-size:2.8vw;
-    margin:7% auto 4%;
-  }
-  .container h3{
-    font-size:4vw;
-    margin:0 0 5%;
-  }
-  .form-container h2, .form-container hr{
-    display:none;
-  }
-  form div{
-    width: 100%;
-    display:flex;
-    flex-direction: column;
-    justify-content: center;
-  }
-  .form-container{
-    width: 100%;
-    justify-content: center;
-  }
-  label{
-    margin:4% 8%;
-  }
-  input{
-    margin:0 auto;
-  }
-  a{ margin: 0 8%;}
-  .form-container .btn-btn-submit{
-    font-size:2.8vw;
-    padding:1.5% 11%;
-  }
-  .form-container .termos {
-    width:62%;
-    margin:0 auto;
-    margin-bottom:5%;
-  }
-  .form-container .termos { font-size:2.8vw;}
-  .form-container p{color:#11111182;}
-  .form-container hr.show-mobile{
-    display:block;
-    margin-bottom:3%;
-  }
-}
