--- src/components/header/index.jsx
@@ -1,11 +1,0 @@
-import "./headerStyle.css"
-
-const Header = () =>{
-  return (
-    <header>
-      <h2>TIL</h2>
-    </header>
-  )
-}
-
-export {Header};
