--- src/components/home/style.css
@@ -1,9 +1,0 @@
-.btn-user{
-    width: 19vw;
-    borer: none;
-    border: none;
-    outline: none;
-    border-radius: 10px;
-    padding: 2% 3%;
-    text-align: right;
-}
