+++ src/login-cadastro/components/header/index.jsx
@@ -1,0 +1,11 @@
+import "./style.css"
+
+const Header = () =>{
+  return (
+    <header>
+      <h2>TIL</h2>
+    </header>
+  )
+}
+
+export {Header};
