+++ src/login-cadastro/components/header/style.css
@@ -1,0 +1,17 @@
+header{
+  height:90px;
+  padding:1%;
+  background:#111111;
+  color:#ffffff;
+  display:flex;
+  align-items:center;
+}
+
+h2{
+  font-size:3vw;
+  margin:5%
+}
+
+@media (max-width:1100px){
+  h2{font-size:8vw;}
+}
