+++ src/pages/home/style.css
@@ -1,6 +1,7 @@
-.container-home{
-  display:flex;
-  flex-direction:column;
-  align-items:center;
-  justify-content:center;
+.container-homePage{
+  width:100%;
+  height:100vh;
+  overflow:hidden;
+  background-image:url("background-paginaInicial.png");
+  background-size:cover;
 }
