+++ src/paginaInicial/components/carrossel-sobre/index.jsx
@@ -1,0 +1,11 @@
+
+
+const CarrosselSobre = () =>{
+  return(
+    <div className="owl-carousel container-text">
+      <p>asdhjkhs</p>
+    </div>
+  )
+}
+
+export { CarrosselSobre }
