+++ src/paginaInicial/components/header/style.css
@@ -1,0 +1,52 @@
+header{
+  width:100%;
+  height:90px;
+  padding:1%;
+  background:#111111;
+  color:#ffffff;
+  display:flex;
+  align-items:center;
+}
+.container-header{
+  width:90%;
+  justify-content:space-between;
+  align-items:center;
+  margin:0 auto;
+  display:flex;
+}
+.container-header .menu{
+  width:50%;
+  display:flex; 
+  justify-content: end;
+}
+.container-header .menu ul{
+  width:90%;
+  display:flex;
+  gap:10%
+}
+.container-header .menu ul li{list-style:none;}
+.container-header .menu ul li a{
+  width:100%;
+  display:block;
+  font-size:2vw;
+  color:#FFFFFF;
+}
+.container-header .menu .btn-header{
+  display:flex;
+  align-items:center;
+  gap:25%;
+  background:none;
+  color:#FFFFFF;
+  font-weight:bold;
+  border: 1px solid #FFFFFF;
+  padding: 0 8%;
+}
+
+h2{
+  font-size:3vw;
+  margin:5%
+}
+
+@media (max-width:1100px){
+  h2{font-size:8vw;}
+}
