+++ src/paginaInicial/components/home/style.css
@@ -1,0 +1,11 @@
+.container-home{
+  width:100%;
+  display:flex;
+  flex-direction:column;
+  align-items:center;
+  justify-content:center;
+}
+.container-home .mascote-home{
+  margin-right:700px;
+  height: 100vh;
+}
