+++ src/paginaInicial/components/mascote/index.jsx
@@ -1,0 +1,9 @@
+const Mascote = () =>{
+ return (
+  <div className="container-mascote">
+    <img src="mascote-til.svg" alt="Mascote TIL" />
+  </div>
+ )
+}
+
+export { Mascote }
