+++ src/paginaInicial/components/mascote/style.css
@@ -1,0 +1,4 @@
+.container-mascote{
+  width:40%;
+  height: 100vh:
+}
